Make containers spill items on polymorphing rather than deleting them
This addresses an uncommon but excruciating source of frustration in the
game: the possibility that you'll typo or be careless when zapping
polymorph beams around (usually to polypile), and hit your stash or bag
of holding by accident. In vanilla, this is punished by vanishing all
the (possibly hundreds) of items in the container, and a common response
is just to quit the game. Thus, to remove this frustration, the
container now spills all of its contents instead of destroying them.

Interestingly, this adds an additional way to extract items from a
statue, because polymorphing the statue will spill those items (even if
it turns into another statue).