Fix: trap ammo: make the fake bones pile use the trap's ammo
I realized that the trap ammo was generating a completely separate stack
of arrows/darts/rocks for the early-level fake bones piles. For rocks
this doesn't really matter, but for arrows and darts it does - they
could generate with a different beatitude, enchantment, material, etc,
and wouldn't stack with the rest of the trap's ammo.

Yes, the trap ammo system does allow for a trap to hold multiple stacks
of ammo and fall back to the next one when the first runs out, but 1)
traps don't actually generate like this, 2) it strains belief that the
trap would have run out of its old ammo perfectly for every fake bones
pile, 3) this way it lets players know if the ammo in the trap is
useful.