Add nh.mon_difficulty() to Lua
This came about because I noticed in a test game that the 'wood nymph'
themed room was generated on dungeon level 1, meaning that if nymphs had
generated there, they would have been out-of-depth, which is a larger
problem in general for themed rooms that want to generate specific
monsters: sure you could set the minimum difficulty of such rooms to the
numeric constant or an offset thereof, but what if the difficulty is
modified at some point?

This commit doesn't fix the wood nymph room (that will come in the next
commit); it just provides the framework for it by exposing a new
mon_difficulty() function to Lua.

I also had to externify find_montype() which did the string-to-permonst
lookup, but was previously restricted to sp_lev.c.