Make the "wood nymph room" generate only at appropriate depth
The non-nymph form of the room will take over when the difficulty of a
wood nymph exceeds the current level difficulty. (Note: Wood nymphs can
still technically generate out-of-depth, because level_difficulty()
isn't monster gen difficulty that factors in the player's experience
level. But it's likely that at some point in the future I'll change
monster gen, at least in_mklev, to disregard the player's level.)