Prevent flipped Sokoban levels from generating
Flipped Sokoban is one of the biggest gripes about 3.7 and I haven't had
a ton of luck convincing the devteam to remove it, so I'm just doing it
myself. Basically, Sokoban-lovers dislike it because it "adds more
levels" to sokoban but not in an intellectually interesting way, whereas
Sokoban-haters who are in it just for the prize hate it because they
can't rely on their muscle memory to do the level quickly.

I would prefer to see more Sokoban levels in the game, but haven't
reviewed candidates thoroughly. This is just a stopgap to get rid of
something players find annoying.