Buff Vorpal Blade artifact damage to +1d8
From UnNetHack. I'm sort of loath to make long sword artifacts even more
attractive, but at the same time, it's kind of a meh weapon for a
crowning gift, when compared with the likes of Excalibur and
Stormbringer.

At some point I may revert this, pending a larger scale artifact
overhaul that possibly replaces the neutral crowning gift.