New seduction success formula
This is because it was far too easy to inflate one's stats to the point
where levels, maximum HP and Pw increases, and Constitution could all be
farmed very reliably. 3.6.1 made an overture towards fixing this by
making it impossible to *guarantee* good outcomes, but you could still
make your chances almost guaranteed.

This alters the denominator of the formula, which in vanilla is a
constant 35, to be based on monster level (which in foocubi ranges from
5 to 9): 25 + m_lev*5, making the target vary from 50 to 70 for foocubi.
Seducers in Gehennom get a bonus of increasing this denominator by 20,
because it's the demons' natural environment. (This is assuming the only
seducers will be demons, which is not likely to change. Also, if I ever
add Malcanthet, their level of 49 will blow off this scale, but there
will still be a small chance of a good outcome, which I find amusing.)

I additionally removed the 3.6.1 cap, meaning you are free again to use
helms of brilliance and rings of adornment to raise your chances as much
as possible. (The theoretical maximum here is 50 from having 25
intelligence and charisma, but that requires both a highly enchanted
helm of brilliance and a very highly enchanted ring of adornment, and
isn't likely to happen much.)

Note that the practice of foocubus dancing isn't actually changed by
this: you can still squeeze good outcomes out of the foocubus until it
gets a severe headache. But it's harder to do that now that the chance
of a good outcome is usually going to be below 90%, and some players may
not find it worth it if the chance is under 50%.