Don't prompt for looting container when there is only one candidate
From UnNetHack, but slightly tweaked due to vanilla having changed how
this looting code works at some point. I also had to make a decision on
what to do if a monster is available to do a directional loot on (not
addressed in UnNetHack), and ended up deciding that if a monster is
next to you while standing over only one container, you are still given
the prompt. The alternative would have been making you start to open the
container (possibly triggering traps) before you were allowed to attempt
looting an adjacent monster.