Differentiate nymphs statblock-wise
From Fourk. The idea here is that we have three species of nymphs in the
game, and the only thing that's really different about them is their
names and colors. Ideally they should not just be renamed clones of each
other.

This sets up a hierarchy: wood < water < mountain in terms of threat,
with wood nymphs a bit slower than the player and mountain a bit faster,
and them all having ascending base levels. To compensate for this, their
difficulties also now ascend.

This is not the last thing I plan to do with differentiating nymphs, but
is likely to be the last in 6.0.