Remove outdated MS-DOS port
I'm quite tired of having function names in those MSC files interfere
with my grep results, and this port existing in the outdated directory
indicates the devteam is considering dropping support for it anyway.
XNetHack has never officially supported this port anyway, as I have
never bothered to update the schema files to include new functions, but
with this I'm officially dropping support for it.