Support themed room sub-areas that respect joined = 0
This enables the development of a couple themed room types I want to
make, namely rooms that contain a shop and some non-shop area. Without
this addition, the random corridor generation treats the shop as
joinable and puts extra doors into it, leaving easy back routes out of
the shop. (The game correctly tracks when you escape a shop by leaving
via a back exit, and shops like this sometimes generate in Bustling
Town, but I still don't want to make them official.)

LIMITATION: Themed rooms can be placed anywhere on the map and the
random corridor generation attempts to dig corridors between the
near-facing walls of those rooms. Thus, it is not safe to have a themed
room where an area marked non-joined occupies the entirety of one or
more walls. One point on the outside of each of the four walls must be
joinable; otherwise it may be forced to place doors on the non-joinable
part anyway and cause impossibles.