Fix: segfault when Wizard of Yendor harassment summoned nasties
Introduced in 18a5bac when I changed the formula here to use the
summoner's level instead of the player's, and evidently forgot that
nasty() can be called with no summoner during harassment.

Now, if the summoner is absent, nasty() will use a monster level of 30,
equal to the Wizard's starting level. I went with a constant value here
because it's simpler than duplicating a code check for what the last
incarnation of the Wizard had as a m_lev, and this makes the real Wizard
nominally more dangerous than his harassment summoning.