Fix: themed room got committed in unfinished test mode
Not sure how this happened. I didn't add it to any commit, but something
went wrong when I tried to push my hotfix branch yesterday. Result was
that the dungeon was in test mode for a new unreleased themed room -
below level 3, any given room had a 50% chance of being the new
twin-business themed room, and a 50% chance of being regular.