New themed room: 'Twin Businesses'
This builds on the commit from two days ago where I introduced the
ability to have themed rooms with sub-areas the corridor generator will
shy away from connecting up. Namely, if you want to have shops in a
themed room, you don't want them to generate extra doors leading to the
rest of the level.

This themed room boasts two shops, a weapons and an armor store, that
can generate in a number of different configurations.