Add `readable' Hawaiian shirts
Like t-shirts and aprons, Hawaiian shirts can now be `read' with 'r' in
order to examine their design. Because this describes the imagery on the
shirt rather than actually reading some sort of text (as is the case
with t-shirts and aprons) this does not break illiterate conduct.

Hawaiian shirt designs are described as <design> on <bg>, for example
``hula dancers on a yellow background'' -- using a combination of
two random strings, rather than limiting the description to just one
primary feature of the shirt, emphasizes the `loudness' stereotypical of
aloha shirts, and can lead to some especially tasteless outfits.