Fix: all random doors specified in lua were turning out secret
Bug in the recent refactor for door-generation unification. Random doors
are represented by -1 (aka 0xffff), and secret doors were specified by
having the 0x20 bit set, so any random door would see that flag set and
decide it should be secret.

Fix this by specifically checking for -1, which should give the door
equal probability of being secret or not.