Refactor: overhaul lspo_door() logic
The way this function was set up was leading to some very confusing
door generations and had limitations such as the inability to specify
doors in partially random states (e.g. a locked door but leaving its
trapped state up to the normal random chance).

In this commit, I've overhauled the function so that it now collects all
the available information on what the programmer specified the door as,
then generates the doorstate fully randomly - then applies the
programmer's specifications to overwrite those parts of it, leaving
anything unspecified as random. So now, e.g. doors in themed rooms
that should be randomly iron/noniron based on their depth don't have to
have it specified.

By doing so, it moves the logic for picking a random door state out of
create_door (which was only used by the door-on-wall code, not the
door-at-coordinate code) into lspo_door. (This is the only caller of
create_door.)