Restore detailed errors to 'W'/'T' and 'P'/'R'
Items that could not be put on/worn were excluded entirely at the getobj
stage rather than simply being discouraged, resulting in the only
feedback given to the player being the generic silly_thing message
instead of the more detailed error messages with the specific reason the
item is not allowed.

Ironically these messages were still present if you used the 'wrong'
command (e.g. 'W' to put on a ring) since that test returned,
discouraging but allowing the item, earlier than the other tests which
would have entirely excluded it.