Fix: don't encourage reading 'novelty' readable items
T-shirts and certain other items that can be read for a funny message
were being included in the getobj list for the 'r' command, which meant
that the list of items could end up always being cluttered with useless
entries for e.g. a hero wearing a t-shirt (even one which could not be
read because it was worn under armor).

I think that the vanilla approach of allowing but not listing these
kinds of items is best, but another option would be to test whether the
item can actually be read (in the case of shirts).  The list of readable
items was also incomplete (for example, it didn't include aprons) so it
would also have to be expanded in this case.