Fix "(illegal)" marker for polearm targeting
The "(illegal)" marker was shown inconsistently and/or incorrectly when
targeting polearm attacks. For example, sea creatures were always marked
"(illegal)" even if they were visible and could be targeted. When blind,
monsters that were legal to target, and which would be hit if selected
them, were marked "(illegal)" even though empty squares were not. This
commit makes "(illegal)" show up with more consistency only on squares
which can actually be targeted if selected.