Refine getobj list for identified thiefstones
Thiefstones previously listed every item in open inventory in their
'a'/#rub prompt; this commit pares down the list only to items which the
hero knows or could guess are compatible with the thiefstone iff the
thiefstone is fully identified, and lists every item in inventory
otherwise, similar to how a touchstone's getobj prompt works.

The list will exclude unidentified magical items which belong to a class
that is mostly mundane (such as magical armor), and will include
unidentified mundane items that belong to a class which is magical by
default (such as potions of sickness or scrolls of blank paper). Items
which are compatible with the thiefstone but are excluded from the
prompt can still be selected, but will not be encouraged in the prompt.