Fix: Properly update symbol of disarmed dart/arrow trap
Untrapping an arrow or dart trap would improperly call newsym before
actually deleting the trap, which meant that the symbol would be
'updated' to the same symbol as before (since the trap was not actually
removed yet).  This was not really noticeable when untrapping something
on the same square as the hero, since the symbol was not visible and
would be updated once the hero moved off the square anyway, but
an adjacent trap would retain its symbol after being disarmed until it
was walked over by the hero or a monster, an item was kicked onto the
square, etc.