Fix: Prevent dog from getting stuck on 'treat' food
Pets would get stuck in food stores (and presumably other places) since
the code that established their current goal did not consider whether
they were hungry, while the code that established whether they ate a
particular item of food did.  This meant that pets would get into a loop
of targeting a desireable food item (e.g. tripe ration), moving over to
it, not eating it because they were not hungry, then getting stuck on it
because it was always the most pressing goal in the vicinity according
to the output of dogfood(dog.c).

Note by copperwater: Added a comment describing DOG_SATIATED; it's one
of those constants that's confusing if you don't know what it does.