Nauseate the player when inside or entering an abbatoir
Abbatoirs are sort of meh rooms - decently flavored with the monsters
and objects and blood, but not particularly interesting otherwise, and
not very dangerous for anyone who can fend off mariliths. With this
change, the danger is upped somewhat, since they now induce nausea. This
effect is tied to the room and can happen indefinitely as long as the
player is inside it, with a bonus chance for happening when they first
encounter the room.

I've wanted to add a nausea source in the later game for a while now: it
adds some interesting complexity in terms of confusing, stunning, and
momentarily making the player helpless while they vomit. (As well as
hitting their hunger slightly.)