If player is foodless, treat them as carrying no food
I thought of this a bit later - a foodless player might be carrying
around food for some reason (pet treats, etc) and their god would see
that they were toting around enough food to get to not-hungry status,
and not respond to a prayer when they get Weak. So, the gods now treat
foodless players as always carrying no food, as far as prayer is
concerned.