Carrying the Amulet blocks teleport control
This is from Fourk, and a gentle buff compared to the previous status
quo where all teleportation was impossible with the Amulet. I notice
that when all teleports are blocked, players tend to accumulate a big
pile of teleport scrolls that they end up never using because by the
ascension run they aren't good for anything, so this will give some
incentive to keep that around (an uncontrolled teleport on a level where
the stairs are far apart is likely to save distance on average). Plus,
as noted in the Fourk commit, teleporting out of danger is an important
escape mechanism.

I have restored the 1/3 chance from vanilla that any horizontal teleport
while carrying the Amulet will fail - so teleporting is still not
guaranteed.