Make paper and straw golems susceptible to rotting attacks
If wood golems can be instakilled by these, then surely the much
flimsier paper and straw golems can also be. I submitted an issue on the
vanilla github repo for this but figured I might as well add it here
because I want it here whether or not they take it.