Fix: cockatrice hissing attack always began stoning if non-deaf
Not related to any of the merging. I just noticed it when I was merging
code into mhitm_ad_ston. This is a pretty severe problem, and appeared
to be due entirely to me typing in the wrong logic rather than an
unforeseen side effect of something.

Bring it back to the original intent: if deaf, the hissing attack does
nothing.