Fix: guard against displacing pets into water got reverted
... by the vanilla commit which introduced displacer beasts and genetic
engineers, and this particular diff apparently applied cleanly to
xNetHack so it flew under the radar. The commit message for that vanilla
commit gives no reason as to why it reverted this, and I've emailed PatR
to ask why, but it is certainly not intended behavior in this variant
for you to be able to instadrown pets by displacing them into water.