Fix: missing death reason for self-cancellation as clay golem
This triggered an impossible once hit, since the polyinit mode assumes
that all callers of rehumanize() must set g.killer prior to calling it.
I initially added the same "was deactivated" death that is used when a
gremlin cancels a clay golem hero, but I grew unhappy with it - deaths
aren't usually in the passive voice - and then decided to change it to
"identity theft", because the writing on/in a golem's head is supposedly
a name for it, in some folklore. Plus, it's much funnier this way.

While here, I fixed another bug where self-canceling while polymorphed
into something other than a clay golem, with unchanging, assumed it was
from an amulet (not necessarily true in xNetHack due to polyinit mode)
and printed an amulet-specific message. Now it has a check around it; I
tried to think of a message that would be printed if your unchanging is
innate, but couldn't, so that will just produce no message.