Bump to xNetHack version 5.1.3
I was going to put this as 5.2, but that would mean all the stuff in the
6.0 changelog now has to be counted as part of a 5.2 release instead.
I'm not prepared to do a release right now because I want the next
release to be 6.0, but I still want to get the savebreaking vanilla
merges out there.

Jumping from 5.1.0 to 5.1.3 because of technical reasons on the
hardfought side.