Wished-for thiefstones key to the player's current location
This logically follows from allowing thiefstones to be re-keyed: if it's
desirable for the player NOT to have to learn the intricacies of the
thiefstone coordinate selection algorithm for the purposes of re-keying
them, this also holds for wished thiefstones. Plus, it'd be a bit silly
for a player to have to go through the rigamarole of re-keying for a
thiefstone they wished for.