Revamp blessed gain ability
One balance problem in NetHack is that blessed gain ability is not a
very interesting item: all attributes are raised equally and there's no
choice involved, so there's never any real reason you might want to save
one for later. Also, I think it's rather too powerful: you only really
need about 3-5 of them in most games to get high enough, if not maxed,
scores in all six attributes, perhaps contributing to the issue where
all high-level characters of a given role feel the same.

This makes it so that blessed gain ability raises one attribute by d2
points, with the player able to choose which or let it choose randomly.
With this change, using it becomes more of a decision into how you want
to build your character (there still isn't much reason to save it for
later unless you're not sure of how your build will go, though). For
instance, it allows a Knight who wants to cast spells to target his
low starting intelligence for boosts.

Also in this commit is a refactor of how gain ability randomly picks an
attribute to raise, for uncursed potions and blessed potions if the
player chooses the random option. Previously it made 6 tries to raise a
random attribute, and if all 6 attempts picked a maxed one you were out
of luck. Now it shuffles an array of the six attributes and iterates
through them until a raisable one is found, so it's still random but
doesn't have a chance of failing to do anything.