Fix: Pass arguments to `mhitm_really_poison` in the correct order
Note from copperwater: The underlying problem (and apparent reason I
wasn't getting any compiler warnings from this) was that
mhitm_really_poison was not declared in either extern.h or the top of
uhitm.c. I augmented this commit, which initially only fixed the calls,
with adding the function declaration - static, at the top of uhitm.c.