Fix listing of gray stones in apply prompt
A comment in apply_ok says that gray stones should be listed in the
apply getobj prompt unless they are known to be neither touchstones nor
thiefstones.  However, the way this was implemented relied on the hero
having identified both touchstones and thiefstones -- if a particular
gray stone was identified as e.g. a flint stone, it would still be
listed in the prompt until touchstones and thiefstones had been
identified.

This commit updates apply_ok so that non-applicable stones can be
identified and excluded either through the process of elimination
("touchstones and thiefstones are identified, so this obj must not be
either" -- this was the extant behavior) or by simple identification of
the gray stone as something other than a touchstone or thiefstone.