Remove forced definition of CURSES_GRAPHICS
Attempting to compile xNetHack with TTY as the sole windowing system
would fail, because despite not including the curses windowing system,
CURSES_GRAPHICS would still be defined in config.h as long as the TTY
window system was included.  I'm not sure what the reason for this was;
removing the definition does not prevent compiling xNetHack with curses,
but it does make compilation and linking work correctly when TTY is the
only windowing system.