Express polyinit form in xlogfile
The xlog already used the flags field to indicate whether a game was
polyinit mode or not, but didn't show any distinction between polyinit
into a strong form like an Archon or a weak form like a jackal. Now it
exposes this.

I considered making the xlog field only be present for polyinit mode
games, but then figured that amid efforts in other variants to remove
inscrutable hex flags from the xlog, it's probably more futureproof to
have it present for all games and show "none" for non-polyinit games.