Add the monstercolor system, letting players choose monster colors
This existed as one of the 3.4.3-nao patches, but rather than porting
that patch to the current codebase I started from scratch, because the
system as I saw it had two big drawbacks:
1. Mutated mons[], which is just not a great idea in my book.
2. Colors were only settable via config file, not in-game. I envisioned
   monster colors working like msgtypes or autopickup exceptions, which
   can be modified in-game.

To that end, monstercolors are settable in-game via a submenu in the
"Other settings" section of the options, like autopickup exceptions or
msgtypes, as well as via "MONSTERCOLOR=monname:color" in the config.
Also like those settings, they do not persist through saving and
restoring, but any config lines will be loaded as normal upon restore.

Instead of directly editing mons[], it edits a new variable
'monstercolors' in instance_globals. I thought this might be unnecessary
at first, but then realized that if it changed mons[].mcolor like the
NAO patch, then it would be impossible to count the number of currently
set monstercolors because there would be no default to compare against.

(cherry picked from commit fcb25b7c7a1cf3aed773f5feaa082ca18b78f38d)