Add "Tea Party" Big Room
Via Fourk, this level was created by jonadab. In terms of monsters,
objects and traps it is identical to the Big Room #1 (a plain,
level-wide room), but there are a number of obstacles arranged in the
center of the room. The four outermost ones are enterable via secret
door.

Fourk hardcoded some tea party "guests" - statues - with specific
monster types and fixed places. I generalized that a bit, adding several
more potential places and names, along with a system for hardcoding
specific monster types for certain names. Special thanks goes to K2 and
his daughter Kathryn, an actual hostess of tea parties and therefore, a
great source of names.