Fix: crash when monster threw glass weapon into iron bars
Technically an object materials bug with glass weapons, though I
consider it a latent vanilla bug because there was nothing guarding
against this possibility. It only doesn't happen in vanilla because
there's no item that: 1) monsters will throw/shoot, 2) which doesn't
explicitly bypass iron bars in hits_bars(), and 3) which breaktest() may
break.

The m_throw() function checked for this in one of its two hits_bars
calls, where the thrown object can crash on iron bars at the end of its
flight path rather than the beginning; it did not check for it on the
other one.