Remove eight polearms
This is from Fourk, and has been a long time coming. NetHack has way
more polearms than it knows what to do with. They're a perennial source
of jokes. Ultimately, we do not need 12 different types of them in the
game.

The remaining 4 polearms were selected to keep some amount of variety
(not that there's all that much of it even now): halberds are still the
all-around best one with good damage to all monsters, glaives are weaker
on small but better on large monsters, partisans are middling against
both, and the bec de corbin is better on small monsters. This is also a
nice balance of piercing, slashing, and blunt weapons, for what that's
worth.

Fourk kept the lucern hammer instead of the bec de corbin, while I did
the opposite, because "lucern hammer" is a confusing name that makes it
sound like a hammer you're supposed to melee with, not pound as a
polearm. However, I copied the lucern hammer's damage stats over to the
bec de corbin, as a very small buff to it.

To avoid disrupting the probability balance of polearms vs other
weapons, I melded the probabilities of the deleted polearms into the
remaining ones.