Add SERVERSEED option to obfuscate hash functions on servers
Players with a good guess of what the server time is at the start of
their game (most existing servers show you the server time) were capable
of guessing their ubirthday. Armed with this knowledge, the values
returned by xNetHack's deterministic hash functions could be fully
guessed.

This commit introduces a sysconf option, SERVERSEED, which allows the
operator of a multi-user NetHack system to throw an additional number
into that hash calculation. Provided that this setting is not shared
with the players on that system, they will no longer be capable of
predicting the results of the hash functions.

I also added a random large number to serverseed at the start of a new
game. The effectiveness of this depends on the system, because it will
be all well and good if entropy is coming from /dev/urandom, but systems
that effectively just do srand(ubirthday) are still predictable if
ubirthday is known. Hence why having it be obfuscated in the sysconf is
still desirable.