Change ad-hoc ubirthday hashing into actual hash functions
A number of places in the code (all from vanilla) were still using
ubirthday directly to do various bits of what amounted to deterministic
hashing. The problem, as explained in the previous commit, is that if a
player knows the local time to within a few seconds, they can predict
how the game will "randomly" decide to do these things. In most cases
where ubirthday was still used, this was relatively minor, but it was
significant in the case of selecting a bones pool.

Thus, this commit changes them to use hash functions instead. Uses of
ubirthday where it was actually used to represent the start time of the
game are not affected.