Fix: rn2(0) in new destroy_items code
If the cap on the number of items destroyed ended up being 0 due to low
incoming damage, the reservoir sampling algorithm would end up calling
rn2(0) with no eligible stacks of items.

If the cap is 0, no items are going to get destroyed anyway, so just
return early from the function in that circumstance.