Fix: findgold always picked first gold item in the object chain
Not sure why I initially implemented it like this, perhaps not
considering it enough of a problem to do anything about. Anyway, it was
enough of a problem to do something about, because it meant that the
player could order their inventory chain with the gold items they didn't
want stolen to come last. Thus, this tosses in a quick reservoir
sampling algorithm that picks the gold item from the whole set at
random.