Enchantment patch part 1: wielded rings plus enchant weapon
This is what is ultimately, I guess, an adaptation of L's Scroll of
Enchantment patch. In that patch, enchant weapon and enchant armor are
merged into a single "scroll of enchantment" item which can be used on
weapons, armor, and also rings. I had this on my todo list but have had
some misgivings about it since then; it seemed like rather than
declutter the object list and simplify the game, it would instead make
it more confusing for players, and Tone also pointed out that it would
disrupt a current balance in which most player builds generally wind up
with a late game surplus of enchant weapon scrolls but a shortage of
enchant armor. So I am no longer planning to port the patch over as
initially implemented.

However, there are still some excellent things that can be done with it,
and the first is making ring enchanting easier for the player to do,
with charging scrolls being rare and better used on other things. Hence
this patch with its surprisingly small diff: reading enchant weapon
while wielding a chargeable ring enchants the ring. The amount of
enchantment gained/lost and ring explosion odds are unchanged.

I think the main problem here will be discoverability; it's very
non-obvious that you can now enchant a ring with a scroll that's still
named enchant weapon, and given that it needs to be wielded to be
enchanted, a name change doesn't make much sense. I added a few rumors
hinting at this, and may add more in the future, but other than that the
best plan might be just to feature it prominently in the next release.