Fix: avoid object lookup overriding monster lookup when both matched
This partially resolves the problem where one uses the / command to look
at Death (or types in "Death"), and answering yes to "More info about
'death'?" instead brings up the object lookup for "wand of death", as
name_to_otyp will find that as the object.

When there is both a mon and object match, it will now prompt the player
to specify which type of lookup they want to see. Additionally, I
noticed that the permonst passed to checkfile() is actually never set to
anything besides NULL, so I piggybacked on that to override when the
player is using / to look at a monster; it won't prompt them to specify
which in that case.

This still isn't perfect; it still can't distinguish what a typed
"death" is looking for, but hopefully the amount of such collisions is
small and the player will now be able to disambiguate the two.