Allow thiefstone rubbed on hands to teleport hero
When the player is polymorphed into a gold golem, rubbing the thiefstone
on their hands (by applying it and selecting '-' to rub on the stone)
should teleport them, as it does when it hits the player when thrown.

This doesn't actually modify anything about the way the thiefstone
teleports the player, just applies the same functionality to rubbing on
hands when polymorphed into a gold golem.