Allow wielded thiefstones to teleport gold golems
Thrown thiefstones would teleport gold golems that they hit back to
their stash location, but wielded thiefstones did not.  This brings
wielded behavior into alignment with thrown behavior.

Also adjusts the way monsters are teleported by a thiefstone, so that
they are transported to the exact keyed coordinates -- a monster will
only be sent to a nearby (enexto) position instead if the keyed location
is not a valid destination spot.